North west corner optimizer cleanup

See merge request !12

+++ src/main/java/de/uni_oldenburg/transport/Solution.java
@@ -105,8 +105,8 @@ public class Solution {
 		return truckTours;
 	}
 
-	public int getTotalConsumption() {
-		int consumption = 0;
+	public double getTotalConsumption() {
+		double consumption = 0;
 		for (Tour tour : truckTours) {
 			consumption += tour.getConsumption();
 		}
